separate skkserv test
