+++ AquaSKK.xcodeproj/xcshareddata/xcschemes/All.App.xcscheme
@@ -51,6 +51,19 @@
             reference = "container:Tests/Integration.xctestplan">
          </TestPlanReference>
       </TestPlans>
+      <Testables>
+         <TestableReference
+            skipped = "NO"
+            parallelizable = "YES">
+            <BuildableReference
+               BuildableIdentifier = "primary"
+               BlueprintIdentifier = "13B93F0A2D908A24008D80C9"
+               BuildableName = "SKKServTests.xctest"
+               BlueprintName = "SKKServTests"
+               ReferencedContainer = "container:AquaSKK.xcodeproj">
+            </BuildableReference>
+         </TestableReference>
+      </Testables>
    </TestAction>
    <LaunchAction
       buildConfiguration = "Debug"
