+++ Tests/CI.xctestplan
@@ -15,8 +15,16 @@
     {
       "target" : {
         "containerPath" : "container:AquaSKK.xcodeproj",
-        "identifier" : "138C69072C94C864009FE153",
-        "name" : "BackendTests"
+        "identifier" : "139B7B6A2C60F68100FE5263",
+        "name" : "UITests"
+      }
+    },
+    {
+      "parallelizable" : false,
+      "target" : {
+        "containerPath" : "container:AquaSKK.xcodeproj",
+        "identifier" : "13B93F0A2D908A24008D80C9",
+        "name" : "SKKServTests"
       }
     },
     {
@@ -29,20 +37,6 @@
     {
       "target" : {
         "containerPath" : "container:AquaSKK.xcodeproj",
-        "identifier" : "1344F6142C5B2C9500D8C850",
-        "name" : "EngineTests"
-      }
-    },
-    {
-      "target" : {
-        "containerPath" : "container:AquaSKK.xcodeproj",
-        "identifier" : "139B7B6A2C60F68100FE5263",
-        "name" : "UITests"
-      }
-    },
-    {
-      "target" : {
-        "containerPath" : "container:AquaSKK.xcodeproj",
         "identifier" : "133834E02C60203900D73AD7",
         "name" : "TyperTests"
       }
@@ -50,8 +44,22 @@
     {
       "target" : {
         "containerPath" : "container:AquaSKK.xcodeproj",
+        "identifier" : "1344F6142C5B2C9500D8C850",
+        "name" : "EngineTests"
+      }
+    },
+    {
+      "target" : {
+        "containerPath" : "container:AquaSKK.xcodeproj",
         "identifier" : "135C3D4D2C5B3E8F00661AE3",
         "name" : "InputTests"
+      }
+    },
+    {
+      "target" : {
+        "containerPath" : "container:AquaSKK.xcodeproj",
+        "identifier" : "138C69072C94C864009FE153",
+        "name" : "BackendTests"
       }
     }
   ],
