+++ xcconfig/Tests-SKKServ.xcconfig
@@ -1,0 +1,12 @@
+//
+//  InputMethodTesting.xcconfig
+//  AquaSKK
+//
+//  Created by mzp on 8/1/24.
+//
+
+// Configuration settings file format documentation can be found at:
+// https://help.apple.com/xcode/#/dev745c5c974
+
+
+OTHER_LDFLAGS = $(inherited)
